requirements.txt
asgiref==3.10.0
Django==5.2.7
gunicorn==23.0.0
packaging==25.0
psycopg-binary==3.2.12
python-dotenv==1.2.1
sqlparse==0.5.3
whitenoise==6.11.0
